requirements.txt
fastapi
uvicorn
sentence-transformers
psycopg2-binary
pgvector
requests
pydantic
